refactor(webhook): decouple dynamic label dispatch
